+++ src/core/io.c
@@ -43,7 +43,7 @@ int io_read_file_part_into_buffer(const char *filepath, int localizable, void *b
     return bytes_read;
 }
 
-int io_write_buffer_to_file(const char *filepath, const void *buffer, int size)
+int io_write_buffer_to_file(const char *filepath, const void *buffer, size_t size)
 {
     // Find existing file to overwrite
     const char *cased_file = dir_get_file(filepath, NOT_LOCALIZED);
